user permissions

+++ controllers/contactController.js
@@ -3,17 +3,17 @@ const Contact = require ("../models/contactModel.js");
 
 //desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async(req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
 
 //desc Create new contacts
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     console.log("The request body is :", req.body);
     const { name, email, phone } = req.body;
@@ -22,14 +22,14 @@ const createContact = asyncHandler(async (req, res) => {
         throw new Error("All fields are mandatory!")
     }
     const contact = await Contact.create({
-        name, email, phone
+        name, email, phone, user_id: req.user.id 
     });
     res.status(201).json(contact);
 });
 
 //desc Get a contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 
 const getContact = asyncHandler(async(req, res) => {
     const contact =  await Contact.findById(req.params.id);
@@ -42,12 +42,17 @@ const getContact = asyncHandler(async(req, res) => {
 
 //desc Update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async(req, res) => {
     const contact =  await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found!");
+    }
+
+    if (contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User does not have permission to update")
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, {new : true });
@@ -56,13 +61,18 @@ const updateContact = asyncHandler(async(req, res) => {
 
 //desc Delete Contact
 //@route Delete /api/contacts/:id
-//@access public
+//@access private
 
 const deleteContact = asyncHandler(async(req, res) => {
     const contact =  await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not found!");
+    }
+
+    if (contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User does not have permission to update")
     }
 
     await Contact.deleteOne(contact)
